Add cloud image based qemu builds
Distributions often ship images which are pre-installed and optimized
for cloud usage which is the case for most `qemu` images.  It is far
better to leverage those existing optimized images and put our
components on top of it rather than install from scratch in an ISO
each time.

This will result in faster build times (preventing the lengthy install
process and making it a simple quick boot) and make builds more reliable
indirectly because of that.  It will also ensure that the image will be
more cloud-native with things like serial interface being mounted
in the right place, etc.

This was setup in a way that existing images can continue to live
with these new ones using a -cloudimg prefix, so both can continue
to work should there be people with a preference to build using
ISO images still.

Since some of the images don't ship with Python, the firstboot playbook
has been updated to support installing Python if it is not installed.